Abstract away Hakyuu functionality to GameType
